+++ server/item.hpp
@@ -13,13 +13,15 @@ class Item
 public:
 
     Item() :
-        _database(NULL), _invalid(true),
+        _database(NULL),
+        _invalid(true), _detached(true), _loaded(false), _changed(false),
         _id(std::numeric_limits<uint64_t>::max())
     {}
 
     Item(Database &database, User const &user, Category const &category,
             uint64_t amount, std::string description) :
-        _database(&database), _invalid(false), _detached(true),
+        _database(&database),
+        _invalid(false), _detached(true), _loaded(true), _changed(true),
         _userid(user.id()), _categoryid(category.id()),
         _amount(amount), _description(description)
     {
@@ -32,10 +34,33 @@ public:
 
     bool valid() { return !_invalid; }
     uint64_t id() const { return _id; }
-    time_t create() { return _create; }
-    time_t modify() { return _modify; }
-    uint64_t amount() { return _amount; }
-    std::string description() { return _description; }
+    time_t create() { lazy_load(); return _create; }
+    time_t modify() { lazy_load(); return _modify; }
+    uint64_t amount() { lazy_load(); return _amount; }
+    std::string description() { lazy_load(); return _description; }
+
+    static void destroy(Database &database, uint64_t userid, uint64_t id)
+    {
+        typedef std::tuple<uint64_t, uint64_t> Params;
+        Params params = std::make_tuple(id, userid);
+        database.execute(params,
+                "DELETE FROM `items` WHERE id = ? and userid = ?");
+        uint64_t affected = database.affected_rows();
+        LOG_DEBUG("Affected: %d", affected);
+        log_assert(affected <= 1);
+        if (affected == 0)
+            throw RowDoesNotExists("Can't destroy item, user does not "
+                    "have that item!");
+    }
+
+    static void destroy(Database &database, Item &obj)
+    {
+        log_assert(!obj._invalid && !obj._detached);
+        destroy(database, obj.user().id(), obj.id());
+        obj._detached = true;
+    }
+
+    void destroy() { destroy(*_database, *this); }
 
     void save()
     {
@@ -102,8 +127,9 @@ private:
             time_t create, time_t modify,
             uint64_t userid, uint64_t categoryid,
             uint64_t amount, std::string description) :
-        _database(&database), _detached(false), _id(id),
-        _create(create), _modify(modify),
+        _database(&database),
+        _invalid(false), _detached(false), _loaded(true), _changed(false),
+        _id(id), _create(create), _modify(modify),
         _userid(userid), _categoryid(categoryid),
         _amount(amount), _description(description)
     {}
@@ -149,9 +175,18 @@ private:
                 });
     }
 
+    void lazy_load()
+    {
+        if (_loaded) return;
+        // TODO
+        log_assert(false);
+    }
+
     Database *_database;
     bool _invalid;
     bool _detached;
+    bool _loaded;
+    bool _changed;
 
     uint64_t _id;
     time_t _create;